Update cloudformation-create-ec2.yml
On request of Pierre added alarm and recovery for instance. There are still Route 53 healthchecks to be added. Due to not be able to push to master and test it on the AWS stack, I've developed the code outside on different AWS account. PR -> merge -> deploy -> test loop would take too much time which I don't have from Pierre.