Location History Overlay V1
When an asset marker is clicked, its history from midnight to the current time is displayed as lines and points. When deselected, the GeoJSON layers and sources are removed from the MapLibre map object.

TODO: Improve overall type-safety and robustness, improve UI line and point elements, implement configuration item to specifically allow the overlay to be visible.