Update permissions needed for Roles and Realms page (#1789)
Roles and Realms used the WRITE_USER permission to determine if the fields are read only. WRITE_USER should be used as permission for allowing a user to write its own authenticated user. Changed the permissions for the Roles and Realms page to match the backend requirements.
Also added description of permissions (roles) on the users page to give a bit more information to the user.